+++ packages/sdk-redux/src/reduxSlices/transactionTrackerSlice/ethersError.ts
@@ -1,8 +1,10 @@
-import {ethers, Transaction} from 'ethers';
+import {ethers} from 'ethers';
+
+import {NewTransactionResponse} from './registerNewTransaction';
 
 export type EthersError = Error & {
     code: ethers.errors;
     reason?: 'reverted' | 'replaced' | 'repriced' | 'cancelled';
     cancelled?: boolean;
-    replacement?: {hash: string} & Transaction;
+    replacement?: NewTransactionResponse;
 };